# wal: reduce allocation when encoding int64