before
```
BenchmarkWrite100EntryWithoutBatch     10000        112254 ns/op         233 B/op          6 allocs/op
BenchmarkWrite100EntryBatch10     200000         11131 ns/op         193 B/op          4 allocs/op
BenchmarkWrite100EntryBatch100    500000          3557 ns/op         192 B/op          4 allocs/op
BenchmarkWrite100EntryBatch500   1000000          1823 ns/op         192 B/op          4 allocs/op
BenchmarkWrite100EntryBatch1000  1000000          1750 ns/op         192 B/op          4 allocs/op
BenchmarkWrite1000EntryWithoutBatch    10000        114098 ns/op        1121 B/op          6 allocs/op
BenchmarkWrite1000EntryBatch10    100000         19193 ns/op        1108 B/op          4 allocs/op
BenchmarkWrite1000EntryBatch100   200000          8259 ns/op        1108 B/op          4 allocs/op
BenchmarkWrite1000EntryBatch500   300000          5725 ns/op        1108 B/op          4 allocs/op
BenchmarkWrite1000EntryBatch1000      300000          5659 ns/op        1108 B/op          4 allocs/op
```

after
```
BenchmarkWrite100EntryWithoutBatch	   10000	    107401 ns/op	     217 B/op	       4 allocs/op
BenchmarkWrite100EntryBatch10	  200000	     10985 ns/op	     177 B/op	       2 allocs/op
BenchmarkWrite100EntryBatch100	  500000	      3393 ns/op	     176 B/op	       2 allocs/op
BenchmarkWrite100EntryBatch500	 1000000	      1737 ns/op	     176 B/op	       2 allocs/op
BenchmarkWrite100EntryBatch1000	 1000000	      1515 ns/op	     176 B/op	       2 allocs/op
BenchmarkWrite1000EntryWithoutBatch	   10000	    114547 ns/op	    1105 B/op	       4 allocs/op
BenchmarkWrite1000EntryBatch10	  100000	     19744 ns/op	    1092 B/op	       2 allocs/op
BenchmarkWrite1000EntryBatch100	  200000	      7931 ns/op	    1092 B/op	       2 allocs/op
BenchmarkWrite1000EntryBatch500	  300000	      5268 ns/op	    1092 B/op	       2 allocs/op
BenchmarkWrite1000EntryBatch1000	  300000	      5587 ns/op	    1092 B/op	       2 allocs/op
```

This pr does not change any encoding logic:
http://golang.org/src/encoding/binary/binary.go?s=6251:6315#L220